differentiate between tear-down and stop-playback
tear-down is cleaner, but may require that set-up happen again before playback
can happen another time, as in the case of stopping and resuming a score.

stop-playback stops playback, and leaves the audio context in a state that it
can continue to play.

When the :one-off? option is present, we want the :stop! function to tear-down,
otherwise we just want to stop playback.